^ change adonis-apollo-server to adonis-graphql + @adonisjs/lucid-slugify to slug title for better SEO + adonis-hashids to encode and decode ID when sending to client

+++ database/migrations/1503248427886_token.js
@@ -5,8 +5,8 @@ const Schema = use('Schema')
 class TokensSchema extends Schema {
   up () {
     this.create('tokens', table => {
-      table.increments()
-      table.integer('user_id').unsigned().references('id').inTable('users')
+      table.bigIncrements()
+      table.bigInteger('user_id').unsigned().references('id').inTable('users')
       table.string('token', 40).notNullable().unique()
       table.string('type', 80).notNullable()
       table.boolean('is_revoked').defaultTo(false)
